requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
plotly>=5.17.0
google-cloud-bigquery>=3.12.0
google-generativeai>=0.3.0
db-dtypes